+++ lib/src/enums/log_type.dart
@@ -1,0 +1,15 @@
+/// An enum representing different types of logs.
+enum LogType {
+  network,
+  button,
+  database,
+  ui,
+  api,
+  other,
+  error,
+  warning,
+  success,
+  info,
+  fatal,
+  timestamp,
+}